requirements.txt
apache-airflow[celery,postgres]==3.1.1
pandas==2.1.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
numpy==1.26.2